# api/docs/swagger.yaml
basePath: /api
definitions:
...
  docs.ErrorResponse:
    properties:
      error:
        example: Error message
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  docs.SuccessResponse:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  handlers.CreateFolderRequest:
    properties:
      name:
        type: string
      path:
        type: string
    type: object
...
  handlers.SharedFolderWithPermission:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      creatorUsername:
        description: Additional fields for display
        type: string
      description:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      memberCount:
        type: integer
      name:
        type: string
      permissionLevel:
        type: integer
      storageQuota:
        description: 0 = unlimited
        type: integer
      updatedAt:
        type: string
      usedStorage:
        type: integer
    type: object
host: localhost:3080
info:
  contact:
    email: support@simplecloudvault.io
    name: API Support
  description: |-
    Enterprise-grade cloud file sharing and management API.
    SimpleCloudVault provides comprehensive file management, sharing, and collaboration features.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SimpleCloudVault API
  version: "1.0"
paths:
...
  /folders:
    post:
      consumes:
      - application/json
      description: Create a new folder at the specified path
      parameters:
      - description: Folder creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFolderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Folder created successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Folder already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create folder
      tags:
      - Files
  /folders/{path}:
    delete:
      consumes:
      - application/json
      description: Delete a folder by path. Use force=true to delete non-empty folders.
      parameters:
      - description: Folder path
        in: path
        name: path
        required: true
        type: string
      - description: Force delete non-empty folder
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Folder deleted successfully
          schema:
            $ref: '#/definitions/docs.SuccessResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Folder is not empty
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete folder
      tags:
      - Files
...
  /shared-folders/my:
    get:
      consumes:
      - application/json
      description: Get list of shared folders the current user has access to
      produces:
      - application/json
      responses:
        "200":
          description: List of shared folders
          schema:
            allOf:
            - $ref: '#/definitions/docs.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.SharedFolderWithPermission'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my shared folders
      tags:
      - SharedFolders
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"